moved metrics to dedicated package
