+++ block/balance.go
@@ -10,6 +10,7 @@ import (
 	"github.com/cockroachdb/errors"
 	"github.com/dymensionxyz/dymint/da"
 	"github.com/dymensionxyz/dymint/types"
+	"github.com/dymensionxyz/dymint/types/metrics"
 )
 
 const CheckBalancesInterval = 3 * time.Minute
@@ -29,7 +30,7 @@ func (m *Manager) MonitorBalances(ctx context.Context) error {
 
 			if balances.DA != nil {
 				if amountFloat, errDA := strconv.ParseFloat(balances.DA.Amount.String(), 64); errDA == nil {
-					types.DaLayerBalanceGauge.Set(amountFloat)
+					metrics.DaLayerBalanceGauge.Set(amountFloat)
 				} else {
 					m.logger.Error("Parsing DA balance amount", "error", errDA)
 				}
@@ -37,7 +38,7 @@ func (m *Manager) MonitorBalances(ctx context.Context) error {
 
 			if balances.SL != nil {
 				if amountFloat, errSL := strconv.ParseFloat(balances.SL.Amount.String(), 64); errSL == nil {
-					types.HubLayerBalanceGauge.Set(amountFloat)
+					metrics.HubLayerBalanceGauge.Set(amountFloat)
 				} else {
 					m.logger.Error("Parsing SL balance amount", "error", errSL)
 				}
