+++ block/block_cache.go
@@ -2,6 +2,7 @@ package block
 
 import (
 	"github.com/dymensionxyz/dymint/types"
+	"github.com/dymensionxyz/dymint/types/metrics"
 )
 
 type Cache struct {
@@ -11,12 +12,12 @@ type Cache struct {
 
 func (m *Cache) Add(h uint64, b *types.Block, c *types.Commit, source types.BlockSource) {
 	m.cache[h] = types.CachedBlock{Block: b, Commit: c, Source: source}
-	types.BlockCacheSizeGauge.Set(float64(m.Size()))
+	metrics.BlockCacheSizeGauge.Set(float64(m.Size()))
 }
 
 func (m *Cache) Delete(h uint64) {
 	delete(m.cache, h)
-	types.BlockCacheSizeGauge.Set(float64(m.Size()))
+	metrics.BlockCacheSizeGauge.Set(float64(m.Size()))
 }
 
 func (m *Cache) Get(h uint64) (types.CachedBlock, bool) {
