+++ block/submit.go
@@ -14,6 +14,7 @@ import (
 	"github.com/dymensionxyz/dymint/da"
 	"github.com/dymensionxyz/dymint/settlement"
 	"github.com/dymensionxyz/dymint/types"
+	"github.com/dymensionxyz/dymint/types/metrics"
 	uchannel "github.com/dymensionxyz/dymint/utils/channel"
 )
 
@@ -181,13 +182,12 @@ func (m *Manager) CreateAndSubmitBatch(maxSizeBytes uint64, lastBatch bool) (*ty
 	if lastBatch && b.EndHeight() == endHeightInclusive {
 		b.LastBatch = true
 	}
-
 	m.logger.Info("Created batch.", "start height", startHeight, "end height", endHeightInclusive, "size", b.SizeBytes(), "last batch", b.LastBatch)
+
 	if err := m.SubmitBatch(b); err != nil {
 		return nil, fmt.Errorf("submit batch: %w", err)
 	}
 
-	types.LastBatchSubmittedBytes.Set(float64(b.SizeBytes()))
 	return b, nil
 }
 
@@ -257,11 +257,13 @@ func (m *Manager) SubmitBatch(batch *types.Batch) error {
 
 	m.logger.Info("Submitted batch to SL.", "start height", batch.StartHeight(), "end height", batch.EndHeight())
 
-	types.RollappHubHeightGauge.Set(float64(batch.EndHeight()))
-	m.LastSettlementHeight.Store(batch.EndHeight())
-
 	// update last submitted block time with batch last block (used to calculate max skew time)
+	m.LastSettlementHeight.Store(batch.EndHeight())
 	m.LastSubmissionTime.Store(time.Now().UTC().UnixNano())
+
+	// update metrics
+	metrics.RollappHubHeightGauge.Set(float64(batch.EndHeight()))
+	metrics.LastBatchSubmittedBytes.Set(float64(batch.SizeBytes()))
 
 	return err
 }
@@ -373,7 +375,7 @@ func (m *Manager) isLastBatchRecent(maxBatchSubmitTime time.Duration) bool {
 }
 
 func UpdateBatchSubmissionGauges(skewBytes uint64, skewBlocks uint64, skewTime time.Duration) {
-	types.RollappPendingSubmissionsBytes.Set(float64(skewBytes))
-	types.RollappPendingSubmissionsBlocks.Set(float64(skewBlocks))
-	types.RollappPendingSubmissionsSkewTimeMinutes.Set(float64(skewTime.Minutes()))
+	metrics.RollappPendingSubmissionsBytes.Set(float64(skewBytes))
+	metrics.RollappPendingSubmissionsBlocks.Set(float64(skewBlocks))
+	metrics.RollappPendingSubmissionsSkewTimeMinutes.Set(float64(skewTime.Minutes()))
 }
