+++ types/metrics/metrics.go
@@ -1,4 +1,4 @@
-package types
+package metrics
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
